+++ src/mobi/hsz/idea/gitignore/IgnoreManager.java
@@ -363,7 +363,7 @@ public class IgnoreManager extends AbstractProjectComponent implements DumbAware
                     if (workingDirectory == null || !Utils.isUnder(file, workingDirectory)) {
                         continue;
                     }
-                    relativePath = StringUtil.trimStart(file.getPath(), workingDirectory.getPath());
+                    relativePath = Utils.getRelativePath(workingDirectory, file);
                 } else {
                     final VirtualFile vcsRoot = getVcsRootFor(file);
                     if (vcsRoot != null && !Utils.isUnder(entryFile, vcsRoot)) {
@@ -372,15 +372,19 @@ public class IgnoreManager extends AbstractProjectComponent implements DumbAware
                         }
                     }
 
-                    final String parentPath = !Utils.isInProject(entryFile, project) &&
-                            project.getBasePath() != null ? project.getBasePath() : entryFile.getParent().getPath();
-                    if (!StringUtil.startsWith(file.getPath(), parentPath) &&
+                    final VirtualFile projectDir = project.getBaseDir();
+                    final VirtualFile parent = !Utils.isInProject(entryFile, project) && projectDir != null
+                            ? projectDir : entryFile.getParent();
+                    if (!StringUtil.startsWith(file.getPath(), parent.getPath()) &&
                             !ExternalIndexableSetContributor.getAdditionalFiles(project).contains(entryFile)) {
                         continue;
                     }
-                    relativePath = StringUtil.trimStart(file.getPath(), parentPath);
+                    relativePath = Utils.getRelativePath(parent, file);
                 }
 
+                if (relativePath == null) {
+                    continue;
+                }
                 relativePath = StringUtil.trimEnd(StringUtil.trimStart(relativePath, "/"), "/");
                 if (StringUtil.isEmpty(relativePath)) {
                     continue;
@@ -471,8 +475,8 @@ public class IgnoreManager extends AbstractProjectComponent implements DumbAware
      * @return file is ignored and tracked
      */
     public boolean isFileTracked(@NotNull final VirtualFile file) {
-        return settings.isInformTrackedIgnored() && !notConfirmedIgnoredFiles.contains(file) &&
-                !confirmedIgnoredFiles.isEmpty() && confirmedIgnoredFiles.containsKey(file);
+        return !notConfirmedIgnoredFiles.contains(file) && !confirmedIgnoredFiles.isEmpty() &&
+                confirmedIgnoredFiles.containsKey(file);
     }
 
     /**
@@ -587,16 +591,6 @@ public class IgnoreManager extends AbstractProjectComponent implements DumbAware
         }
     }
 
-    /**
-     * Returns tracked and ignored files stored in {@link #confirmedIgnoredFiles}.
-     *
-     * @return tracked and ignored files map
-     */
-    @NotNull
-    public ConcurrentMap<VirtualFile, VcsRoot> getConfirmedIgnoredFiles() {
-        return confirmedIgnoredFiles;
-    }
-
     /** {@link Runnable} implementation to rebuild {@link #confirmedIgnoredFiles}. */
     class RefreshTrackedIgnoredRunnable implements Runnable, IgnoreManager.RefreshTrackedIgnoredListener {
         /** Default {@link Runnable} run method that invokes rebuilding with bus event propagating. */
@@ -617,10 +611,6 @@ public class IgnoreManager extends AbstractProjectComponent implements DumbAware
          * @param silent propagate {@link IgnoreManager.TrackedIgnoredListener#TRACKED_IGNORED} event
          */
         public void run(boolean silent) {
-            if (!settings.isInformTrackedIgnored()) {
-                return;
-            }
-
             final ConcurrentMap<VirtualFile, VcsRoot> result = ContainerUtil.newConcurrentMap();
             for (VcsRoot vcsRoot : vcsRoots) {
                 if (!(vcsRoot.getVcs() instanceof GitVcs) || vcsRoot.getPath() == null) {